Refactor key retrieval to search and look at attributes
Refactor key retrieval to search for the key if it's relative and to
match usernames also by key attributes if no key file matching the
username exist.

The side effect of this change is that paths sent to pass are always
sanitized, i.e. no ./foo/user will be passed to pass but foo/user.